Fix problems in the social media tab
The images now take half the horizontal space.
Image height is capped to ensure the caption text is always shorter.
More documentation on the css-tricks used to hide the showmore-button.
Line height and the main colors are now put in css variables.
Added html class shadow to apply the same box shadow to elements while
staying DRY. Updated ajankohtaista css class names to be in line with
the project naming conventions.