Centerbox content top and bottom margins
The margin-bottom of the last paragraphs was removed to make white space
at the end of the page same as on the top and sides.
Content images are now embedded in paragrapsh to ensure correct spacing
and use a class instead of an id, because there aren't unique.
Removed css rules for ul and li elements, which the project no longer
contains.